+++ priv/repo/migrations/20200108032156_create_todos.exs
@@ -5,7 +5,7 @@ defmodule LiveViewTodos.Repo.Migrations.CreateTodos do
     create table(:todos) do
       add :title, :string
       add :done, :boolean, default: false, null: false
-      add :pos, :float, null: false
+      add :pos, :integer, null: false
 
       timestamps()
     end
